fix(button): remove reduced padding for icons
When a button has an icon, we used to reduce padding on the side where
the icon is to make it look more balanced. Unfortunately, this only
really has the desired effect for icons with a lot of whitespace (e.g.
checkmarks), and looks broken for more "busy" icons (e.g. file upload).

This commit removes special padding treatment for icons. The padding is
always the same, whether there is an icon or not. This looks much better
for the more busy icons, and still pretty good for icons with more
whitespace.