feat(button): handle custom slot-based icons
PrimeVue’s <Button> component only automatically positions built-in string-based icons (e.g., icon='pi pi-check'). It does not handle positioning for custom slot-based icons, since it assumes it is controlled manually - which is why we need to manually apply an order style to our custom icon component and pass it through the slot.