feat(inputGroup, inputGroupAddon): simplify + rework on current requirements
The previous versions of the input group and related components was
styled to look like one big input field containing both the actual input
as well as any addons. As far as I can tell this is one of the few
places where we implemented a design that we didn't actually use
anywhere. While using the input group in this way works for that use
case, we now have some requirements that would be very hard to implement
while also keeping the current functionality.

Since we didn't use the input group in its current form, I decided to
start from scratch:

- Removed the current functionality
  - It doesn't seem to be used anywhere
  - Currency signs and other units should be implemented with
    InputNumber instead of combining InputText + InputGroup
  - InputGroups are quite flexible and pretending it's all one big input
    gets difficult to support quickly
- The InputGroup and InputGroupAddon are super simple now, basically
  just a flex container
  - Works for the only use case we have right now (input + button group)
  - If we need more in the future we can add them
  - I wouldn't try to anticipate every possible combination of
    components in the group and rather just extend implementation as
    as needed